updated filestructure: * removed test folders * moved some component folders into _components
